+++ generators/embeddable-graphic/templates/gulp/tasks/aws.js
@@ -3,10 +3,18 @@
 const confirm = require('gulp-confirm');
 const rename = require('gulp-rename');
 const awspublish = require('gulp-awspublish');
+const cloudfront = require('gulp-cloudfront-invalidate-aws-publish');
 const gulp = require('gulp');
 
 const awsJson = require('../../aws.json');
 const meta = require('../../meta.json');
+
+const cfSettings = {
+  distribution: 'E3QK9W6AW5LAVH',
+  accessKeyId: awsJson.accessKeyId,
+  secretAccessKey: awsJson.secretAccessKey,
+  wait: false,
+};
 
 
 module.exports = () => {
@@ -23,5 +31,6 @@ module.exports = () => {
     }))
     .pipe(publisher.publish({}, { force: false }))
     .pipe(publisher.cache())
+    .pipe(cloudfront(cfSettings))
     .pipe(awspublish.reporter());
 };